% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex.R
\name{find_packages}
\alias{find_packages}
\title{Find missing LaTeX packages from a LaTeX log file}
\usage{
find_packages(log, text = readLines(log), quiet = FALSE)
}
\arguments{
\item{log}{Path to the LaTeX log file (typically named \file{*.log}).}

\item{text}{A character vector of the error log (read from the file provided
by the \code{log} argument by default).}

\item{quiet}{Whether to suppress messages when finding packages.}
}
\value{
A character vector of LaTeX package names.
}
\description{
Analyze the error messages in a LaTeX log file to figure out the names of
missing LaTeX packages that caused the errors. These packages can be
installed via \code{\link{tlmgr_install}()}. Searching for missing packages
is based on \command{tlmgr search --global --file}.
}
