% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_tinytex}
\alias{install_tinytex}
\title{Install TinyTeX}
\usage{
install_tinytex()
}
\description{
Download the installation script from \url{https://github.com/yihui/tinytex}
according to the platform (Unix or Windows), and execute it to install
TinyTeX.
}
\references{
TinyTeX homepage: \url{https://yihui.name/tinytex/}.
}
