% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{is_tinytex}
\alias{is_tinytex}
\title{Check if the LaTeX installation is TinyTeX}
\usage{
is_tinytex()
}
\value{
A logical value indicating if the LaTeX installation is TinyTeX.
}
\description{
First find the root directory of the installation via
\code{\link{tinytex_root}()}. Then check if the directory name is
\code{"tinytex"} (case-insensitive). If not, further check if the first line
of the file \file{texmf-dist/web2c/fmtutil.cnf} under the directory contains
\code{"TinyTeX"} or \code{".TinyTeX"}. If the binary version of TinyTeX was
installed, \file{fmtutil.cnf} should contain a line like \samp{Generated by
*/TinyTeX/bin/x86_64-darwin/tlmgr on Thu Sep 17 07:13:28 2020}.
}
\examples{
tinytex::is_tinytex()
}
