% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jup_pkgs.R
\docType{data}
\name{jup_pkgs}
\alias{jup_pkgs}
\title{Additional tinytex packages needed for rendering of PDFs in Jupyter}
\format{
A vector with 16 package names:
\describe{
  \item{adjustbox}{}
  \item{caption}{}
  \item{collectbox}{}
  \item{enumitem}{}
  \item{environ}{}
  \item{eurosym}{}
  \item{jknapltx}{}
  \item{parskip}{}
  \item{pgf}{}
  \item{rsfs}{}
  \item{tcolorbox}{}
  \item{titling}{}
  \item{trimspaces}{}
  \item{ucs}{}
  \item{ulem}{}
  \item{upquote}{}
}
}
\source{
\url{https://github.com/yihui/tinytex/issues/228}
}
\usage{
jup_pkgs
}
\description{
These 16 TexLive packages are required to allow for rendering of PDFs in
Jupyter when tinytex is used to install LaTeX.
}
\keyword{datasets}
