% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex.R
\name{parse_install}
\alias{parse_install}
\title{Parse the LaTeX log and install missing LaTeX packages if possible}
\usage{
parse_install(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link{parse_packages}()}.}
}
\description{
This is a helper function that combines \code{\link{parse_packages}()} and
\code{\link{tlmgr_install}()}.
}
