% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{tinytex-package}
\alias{tinytex}
\alias{tinytex-package}
\title{tinytex: Helper Functions to Install and Maintain TeX Live, and Compile LaTeX Documents}
\description{
Helper functions to install and maintain the 'LaTeX' distribution named 'TinyTeX' (\url{https://yihui.org/tinytex/}), a lightweight, cross-platform, portable, and easy-to-maintain version of 'TeX Live'. This package also contains helper functions to compile 'LaTeX' documents, and install missing 'LaTeX' packages automatically.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rstudio/tinytex}
  \item Report bugs at \url{https://github.com/rstudio/tinytex/issues}
}

}
\author{
\strong{Maintainer}: Yihui Xie \email{xie@yihui.name} (\href{https://orcid.org/0000-0003-0645-5666}{ORCID}) [copyright holder]

Other contributors:
\itemize{
  \item Posit Software, PBC [copyright holder, funder]
  \item Christophe Dervieux (\href{https://orcid.org/0000-0003-4474-2498}{ORCID}) [contributor]
  \item Devon Ryan \email{dpryan79@gmail.com} (\href{https://orcid.org/0000-0002-8549-0971}{ORCID}) [contributor]
  \item Ethan Heinzen [contributor]
  \item Fernando Cagua [contributor]
}

}
\keyword{internal}
