% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tlmgr.R
\name{tl_sizes}
\alias{tl_sizes}
\title{Sizes of LaTeX packages in TeX Live}
\usage{
tl_sizes(show_total = TRUE)
}
\arguments{
\item{show_total}{Whether to show the total size.}
}
\value{
A data frame of three columns: \code{package} is the package names,
  \code{size} is the sizes in bytes, and \code{size_h} is the human-readable
  version of sizes.
}
\description{
Use the command \command{tlmgr info --list --only-installed} to obtain the
sizes of installed LaTeX packages.
}
